LPS-75708 Remove usages, note that while fetching by nodeId and title could return a different wiki page both pages should have the same resource prim key which is what we're che checking permissions on.

+++ modules/apps/collaboration/wiki/wiki-service/src/main/java/com/liferay/wiki/service/impl/WikiPageServiceImpl.java
@@ -35,6 +35,7 @@ import com.liferay.portal.kernel.util.URLCodec;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.spring.extender.service.ServiceReference;
 import com.liferay.rss.util.RSSUtil;
+import com.liferay.subscription.service.SubscriptionLocalService;
 import com.liferay.wiki.configuration.WikiGroupServiceOverriddenConfiguration;
 import com.liferay.wiki.constants.WikiConstants;
 import com.liferay.wiki.constants.WikiPortletKeys;
@@ -184,8 +185,10 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 			ServiceContext serviceContext)
 		throws PortalException {
 
+		WikiPage page = wikiPageLocalService.getPage(nodeId, title, null);
+
 		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.DELETE);
+			getPermissionChecker(), page, ActionKeys.DELETE);
 
 		WikiNodePermissionChecker.check(
 			getPermissionChecker(), nodeId, ActionKeys.ADD_PAGE);
@@ -209,18 +212,32 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 
 	@Override
 	public void deletePage(long nodeId, String title) throws PortalException {
-		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.DELETE);
+		WikiPage page = wikiPagePersistence.fetchByN_T_H_First(
+			nodeId, title, true, null);
 
-		wikiPageLocalService.deletePage(nodeId, title);
+		if (page == null) {
+			return;
+		}
+
+		WikiPagePermissionChecker.check(
+			getPermissionChecker(), page, ActionKeys.DELETE);
+
+		wikiPageLocalService.deletePage(page);
 	}
 
 	@Override
 	public void deletePageAttachment(long nodeId, String title, String fileName)
 		throws PortalException {
 
+		WikiPage page = wikiPagePersistence.fetchByN_T_H_First(
+			nodeId, title, true, null);
+
+		if (page == null) {
+			return;
+		}
+
 		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.DELETE);
+			getPermissionChecker(), page, ActionKeys.DELETE);
 
 		wikiPageLocalService.deletePageAttachment(nodeId, title, fileName);
 	}
@@ -229,8 +246,15 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 	public void deletePageAttachments(long nodeId, String title)
 		throws PortalException {
 
+		WikiPage page = wikiPagePersistence.findByN_T_First(
+			nodeId, title, null);
+
+		if (page == null) {
+			return;
+		}
+
 		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.DELETE);
+			getPermissionChecker(), page, ActionKeys.DELETE);
 
 		wikiPageLocalService.deletePageAttachments(nodeId, title);
 	}
@@ -253,8 +277,11 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 	public void deleteTrashPageAttachments(long nodeId, String title)
 		throws PortalException {
 
+		WikiPage page = wikiPagePersistence.findByN_T_First(
+			nodeId, title, null);
+
 		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.DELETE);
+			getPermissionChecker(), page, ActionKeys.DELETE);
 
 		wikiPageLocalService.deleteTrashPageAttachments(nodeId, title);
 	}
@@ -263,10 +290,12 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 	public void discardDraft(long nodeId, String title, double version)
 		throws PortalException {
 
-		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, version, ActionKeys.DELETE);
+		WikiPage page = wikiPagePersistence.findByN_T_V(nodeId, title, version);
 
-		wikiPageLocalService.discardDraft(nodeId, title, version);
+		WikiPagePermissionChecker.check(
+			getPermissionChecker(), page, ActionKeys.DELETE);
+
+		wikiPagePersistence.remove(page);
 	}
 
 	@Override
@@ -300,10 +329,12 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 	public WikiPage getDraftPage(long nodeId, String title)
 		throws PortalException {
 
-		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.VIEW);
+		WikiPage page = wikiPageLocalService.getDraftPage(nodeId, title);
 
-		return wikiPageLocalService.getDraftPage(nodeId, title);
+		WikiPagePermissionChecker.check(
+			getPermissionChecker(), page, ActionKeys.VIEW);
+
+		return page;
 	}
 
 	@Override
@@ -549,8 +580,16 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 			String attachmentURLPrefix, Locale locale)
 		throws PortalException {
 
-		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.VIEW);
+		WikiPage page = wikiPageLocalService.fetchPage(nodeId, title);
+
+		if (page == null) {
+			WikiNodePermissionChecker.check(
+				getPermissionChecker(), nodeId, ActionKeys.VIEW);
+		}
+		else {
+			WikiPagePermissionChecker.check(
+				getPermissionChecker(), page, ActionKeys.VIEW);
+		}
 
 		List<WikiPage> pages = wikiPageLocalService.getPages(
 			nodeId, title, 0, max, new PageCreateDateComparator(true));
@@ -609,8 +648,10 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 			long nodeId, String title, String fileName)
 		throws PortalException {
 
+		WikiPage page = wikiPageLocalService.getPage(nodeId, title);
+
 		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.DELETE);
+			getPermissionChecker(), page, ActionKeys.DELETE);
 
 		return wikiPageLocalService.movePageAttachmentToTrash(
 			getUserId(), nodeId, title, fileName);
@@ -620,21 +661,29 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 	public WikiPage movePageToTrash(long nodeId, String title)
 		throws PortalException {
 
-		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.DELETE);
+		WikiPage page = wikiPagePersistence.fetchByN_T_H_First(
+			nodeId, title, true, null);
 
-		return wikiPageLocalService.movePageToTrash(getUserId(), nodeId, title);
+		if (page == null) {
+			return null;
+		}
+
+		WikiPagePermissionChecker.check(
+			getPermissionChecker(), page, ActionKeys.DELETE);
+
+		return wikiPageLocalService.movePageToTrash(getUserId(), page);
 	}
 
 	@Override
 	public WikiPage movePageToTrash(long nodeId, String title, double version)
 		throws PortalException {
 
-		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, version, ActionKeys.DELETE);
+		WikiPage page = wikiPagePersistence.findByN_T_V(nodeId, title, version);
 
-		return wikiPageLocalService.movePageToTrash(
-			getUserId(), nodeId, title, version);
+		WikiPagePermissionChecker.check(
+			getPermissionChecker(), page, ActionKeys.DELETE);
+
+		return wikiPageLocalService.movePageToTrash(getUserId(), page);
 	}
 
 	@Override
@@ -643,8 +692,10 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 			ServiceContext serviceContext)
 		throws PortalException {
 
+		WikiPage page = wikiPageLocalService.fetchPage(nodeId, title);
+
 		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.DELETE);
+			getPermissionChecker(), page, ActionKeys.DELETE);
 
 		WikiNodePermissionChecker.check(
 			getPermissionChecker(), nodeId, ActionKeys.ADD_PAGE);
@@ -683,8 +734,10 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 			ServiceContext serviceContext)
 		throws PortalException {
 
+		WikiPage page = wikiPageLocalService.getPage(nodeId, title, version);
+
 		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.UPDATE);
+			getPermissionChecker(), page, ActionKeys.UPDATE);
 
 		return wikiPageLocalService.revertPage(
 			getUserId(), nodeId, title, version, serviceContext);
@@ -694,20 +747,27 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 	public void subscribePage(long nodeId, String title)
 		throws PortalException {
 
-		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.SUBSCRIBE);
+		WikiPage page = wikiPageLocalService.getPage(nodeId, title);
 
-		wikiPageLocalService.subscribePage(getUserId(), nodeId, title);
+		WikiPagePermissionChecker.check(
+			getPermissionChecker(), page, ActionKeys.SUBSCRIBE);
+
+		subscriptionLocalService.addSubscription(
+			getUserId(), page.getGroupId(), WikiPage.class.getName(),
+			page.getResourcePrimKey());
 	}
 
 	@Override
 	public void unsubscribePage(long nodeId, String title)
 		throws PortalException {
 
-		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.SUBSCRIBE);
+		WikiPage page = wikiPageLocalService.getPage(nodeId, title);
 
-		wikiPageLocalService.unsubscribePage(getUserId(), nodeId, title);
+		WikiPagePermissionChecker.check(
+			getPermissionChecker(), page, ActionKeys.SUBSCRIBE);
+
+		subscriptionLocalService.deleteSubscription(
+			getUserId(), WikiPage.class.getName(), page.getResourcePrimKey());
 	}
 
 	@Override
@@ -718,8 +778,16 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 			ServiceContext serviceContext)
 		throws PortalException {
 
-		WikiPagePermissionChecker.check(
-			getPermissionChecker(), nodeId, title, ActionKeys.UPDATE);
+		WikiPage page = wikiPageLocalService.fetchPage(nodeId, title);
+
+		if (page == null) {
+			WikiNodePermissionChecker.check(
+				getPermissionChecker(), nodeId, ActionKeys.VIEW);
+		}
+		else {
+			WikiPagePermissionChecker.check(
+				getPermissionChecker(), page, ActionKeys.UPDATE);
+		}
 
 		return wikiPageLocalService.updatePage(
 			getUserId(), nodeId, title, version, content, summary, minorEdit,
@@ -883,6 +951,9 @@ public class WikiPageServiceImpl extends WikiPageServiceBaseImpl {
 
 	@ServiceReference(type = ConfigurationProvider.class)
 	protected ConfigurationProvider configurationProvider;
+
+	@ServiceReference(type = SubscriptionLocalService.class)
+	protected SubscriptionLocalService subscriptionLocalService;
 
 	@ServiceReference(type = WikiEngineRenderer.class)
 	protected WikiEngineRenderer wikiEngineRenderer;
